Split UnsafeStatements tests into their own file
Additionally move the generic tests of various migration subclasses for
different ActiveRecord versions.
